requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
DateTime==5.5
pydantic==2.11.7
environs==14.2.0
SQLAlchemy==2.0.41
alembic==1.16.4
psycopg2-binary==2.9.10
python-jose[cryptography]
passlib==1.7.4
pillow==11.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
pip==25.1.1
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic_core==2.33.2
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing_extensions==4.14.1
typing-inspection==0.4.1
zope.interface==7.2
bcrypt
uuid